Set up nginx docker image
We put in place an initial organisational structure for docker deployment config
for the nLab. The intention is that all use of docker, both locally and in
production, should be by means of the scripts in deploy/docker/.

We add an nginx image, and set things up so that it can serve nLab pages
statically in a way which persists these pages across docker image builds
(mounting the directory deploy/nlab-static as a volume).