Add renderer of nLab page from database
We borrow more or less unchanged some code from the python renderer in
the current nLab codebase for finding the nForum discussion link
corresponding to a given nLab page, but otherwise this is a complete
rewrite (the current nLab was relying on Rails for this step).

We aim for as clean and as readable a design as possible, making use
of templates.

We thoroughly test the renderer. Instructions for this are given in
the README.md file in test/.

We tweak various aspects of the docker setup, including networking, to
get things to work, including being to use the MySQL database from the
python container.