Complete rendering from database to pages viewable via nginx
This commit builds on the previous one to complete the process of
taking an nLab page as understood by the database, rendering it to
HTML, and placing the rendered HTML page in the correct place for it
to be able to be viewed as a static web page via nginx.

This is principally achieved by the new
render_nlab_page_for_viewing.py. This can be viewed in a stand-alone
way, but it is also intended that the function
render_and_place_for_static_viewing will be called after a page edit.

We add a test which curls the nginx server to view the rendered page.

We tweak various aspects of the docker setup to make all of this
possible. In particular, we now use a custom nginx.conf created
dynamically when calling docker_build.sh, and mount the local
directory in which the static nLab lives live as a volume for the
python docker container as well as the nginx one, so that we can work
with it in python.