Add backend server and render page via it if not already rendered
Previously, nginx would serve a rendered nLab as a static file if it
existed, and return a 404 if not. We now add a backend application
server (CherryPy) with, for the moment, a single GET endpoint which
renders the page in the URL of the endpoint and places it in the
correct directory, as well as returning. We tweak the nginx config to
call this backend endpoint rather than return a 404.

We also make various small abstractions and improvements to the code.
In particular, previously, we had relied on a local MySQL client in a
couple of places; we now tweak things to make this unnecessary. We
still rely on having Python available locally for testing.