Add inline latex rendering to source renderer
We extend the mistletoe markdown renderer to handle inline (between
single dollar signs) LaTeX. We do this in a way which is independent
of the choice of compiler, leaving this to the setting of an
environment variable.

We tweak the deployment to add an itex2MML binary to the python docker
container, and configure things so that this latex compiler will be
used in a deployed nLab.